# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Catalog Service API",
    "description": "<h2>CatalogController is an API for all Catalog service business logic.</h2><br>This functionality extends to:<ul><li> Inserting a new catalog(Registration flow)<li> Inserting a new item for a given merchant.<li> Update an existing item's details.<li> Updating a merchant's public details (e.g.Support information)<li> Retrieving a given merchant's entire catalog (Browsing purposes).</ul>",
    "contact": {
      "name": "Eoin Whelan (Farrell)",
      "email": "C00164354@itcarlow.ie"
    },
    "version": "v1"
  },
  "paths": {
    "/Catalog/InsertMerchant": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "InsertMerchant - Registration flow API call. Inserts new catalog with single item.",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntry"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntry"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogEntry"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Catalog/UpdateMerchant": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "UpdateMerchant - Update flow API call. Updates a given merchant with a dictionary of key/value details.",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "Name of the user.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The updated details.",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Catalog/PullCatalog": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "summary": "Pulls a given username's catalog for storefront display.",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "The username.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Catalog/AddListing": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Adds a new product listing to a given merchant's catalog.",
        "parameters": [
          {
            "name": "username",
            "in": "query",
            "description": "The username.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The new item.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Catalog/UpdateListing": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Updates an existing listing with new details.",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "Name of the user.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The updated item.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItem"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItem"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItem"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Catalog/DeleteListing": {
      "post": {
        "tags": [
          "Catalog"
        ],
        "summary": "Deletes an existing listing from a given merchant's catalog.",
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "description": "Name of the user.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "itemId",
            "in": "query",
            "description": "The item identifier.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CatalogEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "nullable": true
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "instagram": {
            "type": "string",
            "nullable": true
          },
          "twitter": {
            "type": "string",
            "nullable": true
          },
          "catalog": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogItem"
            },
            "nullable": true
          },
          "profileImg": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CatalogItem": {
        "required": [
          "isDeleted",
          "itemDesc",
          "itemName",
          "itemPrice",
          "itemQty"
        ],
        "type": "object",
        "properties": {
          "itemName": {
            "type": "string"
          },
          "itemPrice": {
            "type": "number",
            "format": "double"
          },
          "itemQty": {
            "type": "object",
            "additionalProperties": {
              "type": "integer",
              "format": "int32"
            }
          },
          "itemDesc": {
            "type": "string"
          },
          "itemImg": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "isDeleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      }
    }
  }
}